added error-boundary, user-provider, configured router, etc.
