+++ packages/react-scripts/template/src/components/home/Home.js
@@ -1,0 +1,32 @@
+import React from 'react'
+import { css } from '@emotion/core'
+
+import Logo from './Logo'
+import HomeHeader from './HomeHeader'
+
+const App = () => (
+  <div
+    css={css`
+      text-align: center;
+    `}
+  >
+    <HomeHeader>
+      <Logo />
+      <p>
+        Edit <code>src/components/App.js</code> and save to reload.
+      </p>
+      <a
+        css={css`
+          color: #61dafb;
+        `}
+        href="https://www.familysearch.org/frontier/docs/#/"
+        target="_blank"
+        rel="noopener noreferrer"
+      >
+        Learn Frontier
+      </a>
+    </HomeHeader>
+  </div>
+)
+
+export default App
