+++ packages/react-scripts/template/src/components/home/Home.md
