+++ packages/react-scripts/template/src/components/home/HomeHeader.js
