+++ packages/react-scripts/template/src/components/home/Logo.js
