+++ packages/react-scripts/template/src/components/user/UserInfo.js
@@ -1,0 +1,7 @@
+import React from 'react';
+import {useUser} from '@fs/user';
+
+export default function UserInfo(){
+  const user = useUser();
+  return (<div>Hello, {user.displayName}!</div>)
+}
