+++ packages/react-scripts/template/src/index.js
@@ -1,21 +1,23 @@
-/* eslint-disable react/jsx-filename-extension */
-import React from 'react'
+import React, { Suspense } from 'react'
 import ReactDOM from 'react-dom'
-import { Router } from '@reach/router'
-import './index.css'
+import { UserProvider } from '@fs/user'
+import RootErrorBoundary from '@fs/error-boundary'
 import App from './components/App'
 import * as serviceWorker from './serviceWorker'
 
-ReactDOM.render(
-  <Router>
-    <App path="/" />
-  </Router>,
-  document.getElementById('root')
+const FrontierRoot = () => (
+  <RootErrorBoundary>
+    <Suspense>
+      <UserProvider>
+        <App />
+      </UserProvider>
+    </Suspense>
+  </RootErrorBoundary>
 )
+
+ReactDOM.render(<FrontierRoot />, document.getElementById('root'))
 
 // If you want your app to work offline and load faster, you can change
 // unregister() to register() below. Note this comes with some pitfalls.
 // Learn more about service workers: https://bit.ly/CRA-PWA
 serviceWorker.unregister()
-
-/* eslint-enable react/jsx-filename-extension */
